Wywalenie search bara z widoku ksiązek (szukanie ksiażek z headera)
